+++ src/main/java/com/limikju/daangn_market/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.limikju.daangn_market.controller;
+
+public class MemberController {
+}
